# docs/openapi/swagger.json
{
  "openapi" : "3.0.1",
  "info" : {
    "contact" : {
      "email" : "support@swagger.io",
      "name" : "API Support",
      "url" : "http://www.swagger.io/support"
    },
    "description" : "LocalPaas Backend",
    "license" : {
      "name" : "Apache 2.0",
      "url" : "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "termsOfService" : "http://swagger.io/terms/",
    "title" : "LocalPaas Backend",
    "version" : "0.1"
  },
  "servers" : [ {
    "url" : "/v1"
  } ],
  "paths" : {
    "/api-keys/{ID}" : {
      "get" : {
        "description" : "Gets API key details",
        "operationId" : "getAPIKey",
        "parameters" : [ {
          "description" : "s3 storage ID",
          "in" : "path",
          "name" : "ID",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/github_com_localpaas_localpaas_localpaas_app_usecase_apikeyuc_apikeydto.GetAPIKeyResp"
                }
              }
            },
            "description" : "OK"
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/github_com_localpaas_localpaas_localpaas_app_apperrors.ErrorInfo"
                }
              }
            },
            "description" : "Bad Request"
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/github_com_localpaas_localpaas_localpaas_app_apperrors.ErrorInfo"
                }
              }
            },
            "description" : "Internal Server Error"
          }
        },
        "summary" : "Gets API key details",
        "tags" : [ "api_keys" ]
      }
    },
    "/auth/login-options" : {
      "get" : {
        "description" : "Gets login options",
        "operationId" : "getLoginOptions",
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/github_com_localpaas_localpaas_localpaas_app_usecase_sessionuc_sessiondto.GetLoginOptionsResp"
                }
              }
            },
            "description" : "OK"
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/github_com_localpaas_localpaas_localpaas_app_apperrors.ErrorInfo"
                }
              }
            },
            "description" : "Bad Request"
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/github_com_localpaas_localpaas_localpaas_app_apperrors.ErrorInfo"
                }
              }
            },
            "description" : "Internal Server Error"
          }
        },
        "summary" : "Gets login options",
        "tags" : [ "sessions" ]
      }
    },
    "/auth/login-with-api-key" : {
      "post" : {
        "description" : "Login to system with API key",
        "operationId" : "loginWithAPIKey",
        "requestBody" : {
          "content" : {
            "*/*" : {
              "schema" : {
                "$ref" : "#/components/schemas/github_com_localpaas_localpaas_localpaas_app_usecase_sessionuc_sessiondto.LoginWithAPIKeyReq"
              }
            }
          },
          "description" : "request data",
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/github_com_localpaas_localpaas_localpaas_app_usecase_sessionuc_sessiondto.LoginWithAPIKeyResp"
                }
              }
            },
            "description" : "OK"
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/github_com_localpaas_localpaas_localpaas_app_apperrors.ErrorInfo"
                }
              }
            },
            "description" : "Bad Request"
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/github_com_localpaas_localpaas_localpaas_app_apperrors.ErrorInfo"
                }
              }
            },
            "description" : "Internal Server Error"
          }
        },
        "summary" : "Login to system with API key",
        "tags" : [ "sessions" ],
        "x-codegen-request-body-name" : "body"
      }
    },
    "/auth/login-with-passcode" : {
      "post" : {
        "description" : "Login to system with passcode after using password",
        "operationId" : "loginWithPasscode",
        "requestBody" : {
          "content" : {
            "*/*" : {
              "schema" : {
                "$ref" : "#/components/schemas/github_com_localpaas_localpaas_localpaas_app_usecase_sessionuc_sessiondto.LoginWithPasscodeReq"
              }
            }
          },
          "description" : "request data",
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/github_com_localpaas_localpaas_localpaas_app_usecase_sessionuc_sessiondto.LoginWithPasscodeResp"
                }
              }
            },
            "description" : "OK"
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/github_com_localpaas_localpaas_localpaas_app_apperrors.ErrorInfo"
                }
              }
            },
            "description" : "Bad Request"
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/github_com_localpaas_localpaas_localpaas_app_apperrors.ErrorInfo"
                }
              }
            },
            "description" : "Internal Server Error"
          }
        },
        "summary" : "Login to system with passcode after using password",
        "tags" : [ "sessions" ],
        "x-codegen-request-body-name" : "body"
      }
    },
    "/auth/login-with-password" : {
      "post" : {
        "description" : "When you get response's `next_step` with value `NextMfa`, you need to call the API\n`/login-with-passcode` to complete the login process.",
        "operationId" : "loginWithPassword",
        "requestBody" : {
          "content" : {
            "*/*" : {
              "schema" : {
                "$ref" : "#/components/schemas/github_com_localpaas_localpaas_localpaas_app_usecase_sessionuc_sessiondto.LoginWithPasswordReq"
              }
            }
          },
          "description" : "request data",
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/github_com_localpaas_localpaas_localpaas_app_usecase_sessionuc_sessiondto.LoginWithPasswordResp"
                }
              }
            },
            "description" : "OK"
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/github_com_localpaas_localpaas_localpaas_app_apperrors.ErrorInfo"
                }
              }
            },
            "description" : "Bad Request"
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/github_com_localpaas_localpaas_localpaas_app_apperrors.ErrorInfo"
                }
              }
            },
            "description" : "Internal Server Error"
          }
        },
        "summary" : "Login to system with username/password",
        "tags" : [ "sessions" ],
        "x-codegen-request-body-name" : "body"
      }
    },
    "/cluster/nodes" : {
      "get" : {
        "description" : "Lists cluster nodes",
        "operationId" : "listNode",
        "parameters" : [ {
          "description" : "`status=<target>`",
          "in" : "query",
          "name" : "status",
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "`search=<target> (support *)`",
          "in" : "query",
          "name" : "search",
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "`pageOffset=offset`",
          "in" : "query",
          "name" : "pageOffset",
          "schema" : {
            "type" : "integer"
          }
        }, {
          "description" : "`pageLimit=limit`",
          "in" : "query",
          "name" : "pageLimit",
          "schema" : {
            "type" : "integer"
          }
        }, {
          "description" : "`sort=[-]field1|field2...`",
          "in" : "query",
          "name" : "sort",
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/github_com_localpaas_localpaas_localpaas_app_usecase_clusteruc_clusterdto.ListNodeResp"
                }
              }
            },
            "description" : "OK"
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/github_com_localpaas_localpaas_localpaas_app_apperrors.ErrorInfo"
                }
              }
            },
            "description" : "Bad Request"
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/github_com_localpaas_localpaas_localpaas_app_apperrors.ErrorInfo"
                }
              }
            },
            "description" : "Internal Server Error"
          }
        },
        "summary" : "Lists cluster nodes",
        "tags" : [ "cluster_nodes" ]
      }
    },
    "/cluster/nodes/{nodeID}" : {
      "delete" : {
        "description" : "Deletes a node",
        "operationId" : "deleteNode",
        "parameters" : [ {
          "description" : "node ID",
          "in" : "path",
          "name" : "nodeID",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/github_com_localpaas_localpaas_localpaas_app_usecase_clusteruc_clusterdto.DeleteNodeResp"
                }
              }
            },
            "description" : "OK"
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/github_com_localpaas_localpaas_localpaas_app_apperrors.ErrorInfo"
                }
              }
            },
            "description" : "Bad Request"
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/github_com_localpaas_localpaas_localpaas_app_apperrors.ErrorInfo"
                }
              }
            },
            "description" : "Internal Server Error"
          }
        },
        "summary" : "Deletes a node",
        "tags" : [ "cluster_nodes" ]
      },
      "get" : {
        "description" : "Gets node details",
        "operationId" : "getNode",
        "parameters" : [ {
          "description" : "node ID",
          "in" : "path",
          "name" : "nodeID",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/github_com_localpaas_localpaas_localpaas_app_usecase_clusteruc_clusterdto.GetNodeResp"
                }
              }
            },
            "description" : "OK"
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/github_com_localpaas_localpaas_localpaas_app_apperrors.ErrorInfo"
                }
              }
            },
            "description" : "Bad Request"
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/github_com_localpaas_localpaas_localpaas_app_apperrors.ErrorInfo"
                }
              }
            },
            "description" : "Internal Server Error"
          }
        },
        "summary" : "Gets node details",
        "tags" : [ "cluster_nodes" ]
      }
    },
    "/internal/auth/dev-mode-login" : {
      "post" : {
        "description" : "Login user for dev mode. `user_id` params is required.",
        "operationId" : "devModeLogin",
        "parameters" : [ {
          "description" : "user id to login",
          "in" : "query",
          "name" : "user_id",
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/github_com_localpaas_localpaas_localpaas_app_usecase_sessionuc_sessiondto.DevModeLoginResp"
                }
              }
            },
            "description" : "OK"
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/github_com_localpaas_localpaas_localpaas_app_apperrors.ErrorInfo"
                }
              }
            },
            "description" : "Bad Request"
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/github_com_localpaas_localpaas_localpaas_app_apperrors.ErrorInfo"
                }
              }
            },
            "description" : "Internal Server Error"
          }
        },
        "summary" : "Login user for dev mode",
        "tags" : [ "sessions" ]
      }
    },
    "/projects" : {
      "get" : {
        "description" : "Lists projects",
        "operationId" : "listProject",
        "parameters" : [ {
          "description" : "`status=<target>`",
          "in" : "query",
          "name" : "status",
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "`search=<target> (support *)`",
          "in" : "query",
          "name" : "search",
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "`pageOffset=offset`",
          "in" : "query",
          "name" : "pageOffset",
          "schema" : {
            "type" : "integer"
          }
        }, {
          "description" : "`pageLimit=limit`",
          "in" : "query",
          "name" : "pageLimit",
          "schema" : {
            "type" : "integer"
          }
        }, {
          "description" : "`sort=[-]field1|field2...`",
          "in" : "query",
          "name" : "sort",
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/github_com_localpaas_localpaas_localpaas_app_usecase_projectuc_projectdto.ListProjectResp"
                }
              }
            },
            "description" : "OK"
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/github_com_localpaas_localpaas_localpaas_app_apperrors.ErrorInfo"
                }
              }
            },
            "description" : "Bad Request"
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/github_com_localpaas_localpaas_localpaas_app_apperrors.ErrorInfo"
                }
              }
            },
            "description" : "Internal Server Error"
          }
        },
        "summary" : "Lists projects",
        "tags" : [ "projects" ]
      },
      "post" : {
        "description" : "Creates a new project",
        "operationId" : "createProject",
        "requestBody" : {
          "content" : {
            "*/*" : {
              "schema" : {
                "$ref" : "#/components/schemas/github_com_localpaas_localpaas_localpaas_app_usecase_projectuc_projectdto.CreateProjectReq"
              }
            }
          },
          "description" : "request data",
          "required" : true
        },
        "responses" : {
          "201" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/github_com_localpaas_localpaas_localpaas_app_usecase_projectuc_projectdto.CreateProjectResp"
                }
              }
            },
            "description" : "Created"
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/github_com_localpaas_localpaas_localpaas_app_apperrors.ErrorInfo"
                }
              }
            },
            "description" : "Bad Request"
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/github_com_localpaas_localpaas_localpaas_app_apperrors.ErrorInfo"
                }
              }
            },
            "description" : "Internal Server Error"
          }
        },
        "summary" : "Creates a new project",
        "tags" : [ "projects" ],
        "x-codegen-request-body-name" : "body"
      }
    },
    "/projects/base" : {
      "get" : {
        "description" : "Lists projects",
        "operationId" : "listProjectBase",
        "parameters" : [ {
          "description" : "`status=<target>`",
          "in" : "query",
          "name" : "status",
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "`search=<target> (support *)`",
          "in" : "query",
          "name" : "search",
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "`pageOffset=offset`",
          "in" : "query",
          "name" : "pageOffset",
          "schema" : {
            "type" : "integer"
          }
        }, {
          "description" : "`pageLimit=limit`",
          "in" : "query",
          "name" : "pageLimit",
          "schema" : {
            "type" : "integer"
          }
        }, {
          "description" : "`sort=[-]field1|field2...`",
          "in" : "query",
          "name" : "sort",
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/github_com_localpaas_localpaas_localpaas_app_usecase_projectuc_projectdto.ListProjectBaseResp"
                }
              }
            },
            "description" : "OK"
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/github_com_localpaas_localpaas_localpaas_app_apperrors.ErrorInfo"
                }
              }
            },
            "description" : "Bad Request"
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/github_com_localpaas_localpaas_localpaas_app_apperrors.ErrorInfo"
                }
              }
            },
            "description" : "Internal Server Error"
          }
        },
        "summary" : "Lists projects",
        "tags" : [ "projects" ]
      }
    },
    "/projects/{projectID}" : {
      "delete" : {
        "description" : "Deletes a project",
        "operationId" : "deleteProject",
        "parameters" : [ {
          "description" : "project ID",
          "in" : "path",
          "name" : "projectID",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/github_com_localpaas_localpaas_localpaas_app_usecase_projectuc_projectdto.DeleteProjectResp"
                }
              }
            },
            "description" : "OK"
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/github_com_localpaas_localpaas_localpaas_app_apperrors.ErrorInfo"
                }
              }
            },
            "description" : "Bad Request"
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/github_com_localpaas_localpaas_localpaas_app_apperrors.ErrorInfo"
                }
              }
            },
            "description" : "Internal Server Error"
          }
        },
        "summary" : "Deletes a project",
        "tags" : [ "projects" ]
      },
      "get" : {
        "description" : "Gets project details",
        "operationId" : "getProject",
        "parameters" : [ {
          "description" : "project ID",
          "in" : "path",
          "name" : "projectID",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/github_com_localpaas_localpaas_localpaas_app_usecase_projectuc_projectdto.GetProjectResp"
                }
              }
            },
            "description" : "OK"
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/github_com_localpaas_localpaas_localpaas_app_apperrors.ErrorInfo"
                }
              }
            },
            "description" : "Bad Request"
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/github_com_localpaas_localpaas_localpaas_app_apperrors.ErrorInfo"
                }
              }
            },
            "description" : "Internal Server Error"
          }
        },
        "summary" : "Gets project details",
        "tags" : [ "projects" ]
      }
    },
...
    "/sessions" : {
      "delete" : {
        "description" : "Deletes the current user session",
        "operationId" : "deleteSession",
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/github_com_localpaas_localpaas_localpaas_app_usecase_sessionuc_sessiondto.DeleteSessionResp"
                }
              }
            },
            "description" : "OK"
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/github_com_localpaas_localpaas_localpaas_app_apperrors.ErrorInfo"
                }
              }
            },
            "description" : "Bad Request"
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/github_com_localpaas_localpaas_localpaas_app_apperrors.ErrorInfo"
                }
              }
            },
            "description" : "Internal Server Error"
          }
        },
        "summary" : "Deletes the current user session",
        "tags" : [ "sessions" ]
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "github_com_localpaas_localpaas_localpaas_app_basedto.BaseMeta" : {
        "properties" : {
          "code" : {
            "type" : "string"
          },
          "message" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "github_com_localpaas_localpaas_localpaas_app_usecase_apikeyuc_apikeydto.GetAPIKeyResp" : {
        "properties" : {
          "data" : {
            "$ref" : "#/components/schemas/github_com_localpaas_localpaas_localpaas_app_usecase_apikeyuc_apikeydto.APIKeyResp"
          },
          "meta" : {
            "$ref" : "#/components/schemas/github_com_localpaas_localpaas_localpaas_app_basedto.BaseMeta"
          }
        },
        "required" : [ "data", "meta" ],
        "type" : "object"
      },
...
      "github_com_localpaas_localpaas_localpaas_app_usecase_sessionuc_sessiondto.GetLoginOptionsResp" : {
        "properties" : {
          "data" : {
            "$ref" : "#/components/schemas/github_com_localpaas_localpaas_localpaas_app_usecase_sessionuc_sessiondto.LoginOptionsResp"
          },
          "meta" : {
            "$ref" : "#/components/schemas/github_com_localpaas_localpaas_localpaas_app_basedto.BaseMeta"
          }
        },
        "required" : [ "data", "meta" ],
        "type" : "object"
      },
...
    },
    "securitySchemes" : {
      "BasicAuth" : {
        "scheme" : "basic",
        "type" : "http"
      }
    }
  },
...
}