# localpaas_app/interface/api/server/router.go
package server

import (
	"net/http"

	"github.com/gin-gonic/gin"
	swaggoFiles "github.com/swaggo/files"
	swaggoGin "github.com/swaggo/gin-swagger"

	"github.com/localpaas/localpaas/localpaas_app/interface/api/handler/apikeyhandler"
	"github.com/localpaas/localpaas/localpaas_app/interface/api/handler/apphandler"
	"github.com/localpaas/localpaas/localpaas_app/interface/api/handler/authhandler"
	"github.com/localpaas/localpaas/localpaas_app/interface/api/handler/clusterhandler"
	"github.com/localpaas/localpaas/localpaas_app/interface/api/handler/projecthandler"
	"github.com/localpaas/localpaas/localpaas_app/interface/api/handler/s3storagehandler"
	"github.com/localpaas/localpaas/localpaas_app/interface/api/handler/sessionhandler"
	"github.com/localpaas/localpaas/localpaas_app/interface/api/handler/sshkeyhandler"
	"github.com/localpaas/localpaas/localpaas_app/interface/api/handler/userhandler"
)

type HandlerRegistry struct {
	authHandler      *authhandler.AuthHandler
	clusterHandler   *clusterhandler.ClusterHandler
	sessionHandler   *sessionhandler.SessionHandler
	userHandler      *userhandler.UserHandler
	projectHandler   *projecthandler.ProjectHandler
	appHandler       *apphandler.AppHandler
	s3StorageHandler *s3storagehandler.S3StorageHandler
	sshKeyHandler    *sshkeyhandler.SSHKeyHandler
	apiKeyHandler    *apikeyhandler.APIKeyHandler
}

func NewHandlerRegistry(
	authHandler *authhandler.AuthHandler,
	clusterHandler *clusterhandler.ClusterHandler,
	sessionHandler *sessionhandler.SessionHandler,
	userHandler *userhandler.UserHandler,
	projectHandler *projecthandler.ProjectHandler,
	appHandler *apphandler.AppHandler,
	s3StorageHandler *s3storagehandler.S3StorageHandler,
	sshKeyHandler *sshkeyhandler.SSHKeyHandler,
	apiKeyHandler *apikeyhandler.APIKeyHandler,
) *HandlerRegistry {
	return &HandlerRegistry{
		authHandler:      authHandler,
		clusterHandler:   clusterHandler,
		sessionHandler:   sessionHandler,
		userHandler:      userHandler,
		projectHandler:   projectHandler,
		appHandler:       appHandler,
		s3StorageHandler: s3StorageHandler,
		sshKeyHandler:    sshKeyHandler,
		apiKeyHandler:    apiKeyHandler,
	}
}

//nolint:funlen
func (s *HTTPServer) registerRoutes() {
	s.engine.GET("/_/ping", routePing)
	s.engine.NoRoute(routeNotFound)

	// Swagger server
	if !s.config.IsProdEnv() {
		s.engine.Use(StaticServe("/docs", localFile("./docs", false)))
		s.engine.GET("/swagger/*any", swaggoGin.WrapHandler(swaggoFiles.Handler,
			swaggoGin.URL("/docs/openapi/swagger.json")))
	}

	// STATIC FILES
	s.engine.Use(StaticServe("/files/user/photo", localFile(s.config.App.DataPathUserPhoto(), false)))
	// Serve the static files from the "dist-dashboard" directory at the root URL "/"
	s.engine.Use(StaticServe("/", localFile("./dist-dashboard", true)))

	// INTERNAL ROUTES
	basicAuthMdlw := gin.BasicAuth(gin.Accounts{
		s.config.Session.BasicAuth.Username: s.config.Session.BasicAuth.Password,
	})
	v1BasicAuth := s.engine.Group(s.config.HTTPServer.BasePath + "/internal")
	v1BasicAuth.Use(basicAuthMdlw)

	// Dev mode
	if !s.config.IsProdEnv() {
		v1BasicAuth.POST("/auth/dev-mode-login", s.handlerRegistry.sessionHandler.DevModeLogin)
	}

	// PUBLIC ROUTES
	apiV1Group := s.engine.Group(s.config.HTTPServer.BasePath)

	clusterGroup := apiV1Group.Group("/cluster")
	{ // node group
		nodeGroup := clusterGroup.Group("/nodes")
		// Nodes
		nodeGroup.GET("", s.handlerRegistry.clusterHandler.ListNode)
		nodeGroup.GET("/:nodeID", s.handlerRegistry.clusterHandler.GetNode)
		nodeGroup.DELETE("/:nodeID", s.handlerRegistry.clusterHandler.DeleteNode)
	}

	{ // session group
		sessionGroup := apiV1Group.Group("/sessions")
		// User info
		sessionGroup.GET("/me", s.handlerRegistry.sessionHandler.GetMe)
		// Session handling
		sessionGroup.POST("/refresh", s.handlerRegistry.sessionHandler.RefreshSession)
		sessionGroup.DELETE("", s.handlerRegistry.sessionHandler.DeleteSession)
		sessionGroup.POST("/delete-all", s.handlerRegistry.sessionHandler.DeleteAllSessions)
	}

	{ // auth group
		authGroup := apiV1Group.Group("/auth")
		// Login options
		authGroup.GET("/login-options", s.handlerRegistry.sessionHandler.LoginGetOptions)
		// Login with password
		authGroup.POST("/login-with-password", s.handlerRegistry.sessionHandler.LoginWithPassword)
		authGroup.POST("/login-with-passcode", s.handlerRegistry.sessionHandler.LoginWithPasscode)
		// Login with API key
		authGroup.POST("/login-with-api-key", s.handlerRegistry.sessionHandler.LoginWithAPIKey)
	}

	{ // user group
		userGroup := apiV1Group.Group("/users")
		// Get user info
		userGroup.GET("/base", s.handlerRegistry.userHandler.ListUserBase)
		userGroup.GET("/:userID", s.handlerRegistry.userHandler.GetUser)
		userGroup.GET("", s.handlerRegistry.userHandler.ListUser)
		// Password
		userGroup.PATCH("/current/password", s.handlerRegistry.userHandler.UpdateUserPassword)
		// MFA TOTP setup
		userGroup.POST("/current/mfa/totp-begin-setup", s.handlerRegistry.userHandler.BeginMFATotpSetup)
		userGroup.POST("/current/mfa/totp-complete-setup", s.handlerRegistry.userHandler.CompleteMFATotpSetup)
		userGroup.POST("/current/mfa/totp-remove", s.handlerRegistry.userHandler.RemoveMFATotp)
		// Invite & SignUp
		userGroup.POST("/invite", s.handlerRegistry.userHandler.InviteUser)
		userGroup.POST("/signup-begin", s.handlerRegistry.userHandler.BeginUserSignup)
		userGroup.POST("/signup-complete", s.handlerRegistry.userHandler.CompleteUserSignup)

		// API key group
		apiKeyGroup := userGroup.Group("/current/api-keys")
		{
			// Info
			apiKeyGroup.GET("/:ID", s.handlerRegistry.apiKeyHandler.GetAPIKey)
			apiKeyGroup.GET("", s.handlerRegistry.apiKeyHandler.ListAPIKey)
			// Creation & Update
			apiKeyGroup.POST("", s.handlerRegistry.apiKeyHandler.CreateAPIKey)
			apiKeyGroup.DELETE("/:ID", s.handlerRegistry.apiKeyHandler.DeleteAPIKey)
		}
	}

	projectGroup := apiV1Group.Group("/projects")
	{ // project group
		// Project info
		projectGroup.GET("/base", s.handlerRegistry.projectHandler.ListProjectBase)
		projectGroup.GET("/:projectID", s.handlerRegistry.projectHandler.GetProject)
		projectGroup.GET("", s.handlerRegistry.projectHandler.ListProject)
		// Creation & Update
		projectGroup.POST("", s.handlerRegistry.projectHandler.CreateProject)
		projectGroup.DELETE("/:projectID", s.handlerRegistry.projectHandler.DeleteProject)
		// Tags
		projectGroup.POST("/:projectID/tags", s.handlerRegistry.projectHandler.CreateProjectTag)
		projectGroup.POST("/:projectID/tags/delete", s.handlerRegistry.projectHandler.DeleteProjectTags)
		// Settings
		projectGroup.GET("/:projectID/settings", s.handlerRegistry.projectHandler.GetProjectSettings)
		projectGroup.PATCH("/:projectID/settings", s.handlerRegistry.projectHandler.UpdateProjectSettings)
	}

	appGroup := projectGroup.Group("/:projectID/apps")
	{ // app group
		// App info
		appGroup.GET("/base", s.handlerRegistry.appHandler.ListAppBase)
		appGroup.GET("/:appID", s.handlerRegistry.appHandler.GetApp)
		appGroup.GET("", s.handlerRegistry.appHandler.ListApp)
		// Creation & Update
		appGroup.POST("", s.handlerRegistry.appHandler.CreateApp)
		appGroup.DELETE("/:appID", s.handlerRegistry.appHandler.DeleteApp)
		// Tags
		appGroup.POST("/:appID/tags", s.handlerRegistry.appHandler.CreateAppTag)
		appGroup.POST("/:appID/tags/delete", s.handlerRegistry.appHandler.DeleteAppTags)
		// Settings
		appGroup.GET("/:appID/settings", s.handlerRegistry.appHandler.GetAppSettings)
		appGroup.PATCH("/:appID/settings", s.handlerRegistry.appHandler.UpdateAppSettings)
	}

	s3StorageGroup := apiV1Group.Group("/s3-storages")
	{ // s3 storage group
		// Info
		s3StorageGroup.GET("/base", s.handlerRegistry.s3StorageHandler.ListS3StorageBase)
		s3StorageGroup.GET("/:ID", s.handlerRegistry.s3StorageHandler.GetS3Storage)
		s3StorageGroup.GET("", s.handlerRegistry.s3StorageHandler.ListS3Storage)
		// Creation & Update
		s3StorageGroup.POST("", s.handlerRegistry.s3StorageHandler.CreateS3Storage)
		s3StorageGroup.PATCH("/:ID", s.handlerRegistry.s3StorageHandler.UpdateS3Storage)
		s3StorageGroup.DELETE("/:ID", s.handlerRegistry.s3StorageHandler.DeleteS3Storage)
	}

	sshKeyGroup := apiV1Group.Group("/ssh-keys")
	{ // ssh key group
		// Info
		sshKeyGroup.GET("/base", s.handlerRegistry.sshKeyHandler.ListSSHKeyBase)
		sshKeyGroup.GET("/:ID", s.handlerRegistry.sshKeyHandler.GetSSHKey)
		sshKeyGroup.GET("", s.handlerRegistry.sshKeyHandler.ListSSHKey)
		// Creation & Update
		sshKeyGroup.POST("", s.handlerRegistry.sshKeyHandler.CreateSSHKey)
		sshKeyGroup.PATCH("/:ID", s.handlerRegistry.sshKeyHandler.UpdateSSHKey)
		sshKeyGroup.DELETE("/:ID", s.handlerRegistry.sshKeyHandler.DeleteSSHKey)
	}
}

func routePing(c *gin.Context) {
	c.JSON(http.StatusOK, gin.H{"message": "pong"})
}

func routeNotFound(c *gin.Context) {
	c.JSON(http.StatusNotFound, "not found")
}